Change creation ui of Dlang Module and project
With this change the creation of module & project look like other
languages. An extension is added to allow dub to wire itself to the D
Language page (like for gradle and Maven in Java). This allow to have
dub being more decoupled to D core, and to have better UI and better UX.